Move URL routines to examples dir (#309)
* Update blossom-ci.yml (#295)

* Change pip install commands with the correct PyPI package name (#298)

* reorganize

* remove from init

---------
