+++ modulus/utils/generative/__init__.py
@@ -36,16 +36,12 @@ from .utils import (
     get_top_level_function_name,
     is_pickleable,
     is_top_level_function,
-    is_url,
     list_dir_recursively_with_ignore,
-    make_cache_dir_path,
     named_params_and_buffers,
-    open_url,
     params_and_buffers,
     parse_int_list,
     print_module_summary,
     profiled_function,
-    set_cache_dir,
     suppress_tracer_warnings,
     tuple_product,
 )
